+++ noodles-vcf/CHANGELOG.md
@@ -22,6 +22,12 @@
     Change instantiations of `vcf::reader::Builder` to
     `vcf::reader::Builder::default()`.
 
+  * vcf/reader/header: Parse header line by line.
+
+    The header parser can now build a `vcf::Header` by parsing a raw header
+    line by line. This makes it so that it is no longer required to read the
+    entire raw header into memory before parsing.
+
   * vcf/writer/builder: Add `Builder::build_from_writer`.
 
   * vcf/writer/builder: Add a compression method setter.